fix: Disable sendside video suspension for the LipSyncTest.
The `googEnableVideoSuspendBelowMinBitrate` flag is breaking our lipsync
test because sending video is suspended and never reactivated. It could
be due to several things: a problem in our padding termination (their
padding is killed before it reaches our remote bitrate estimator), a
problem in the old GCC implementation, or a problem in their padding
budget calculation.