fix: change selectors and behavior for lock test
The invite modal UX has been changed to add a toggle link for
showing and hiding the fields to edit the room password. The
modal has also been converted to use react. These changes require
modifying LockRoomTest and how it searches for elements and sets
or removes the room password.