fix(mute-test): hover over different remote menu trigger element
Two flaky issues have cropped up. One is the remote menu is not
displaying on all instances when clicking on remote menu trigger.
Admittedly the cause is not completely known, but using the
trigger container makes the menu show more consistently -- this
may be an implementation detail of AtlasKit InlineDialog, which
wraps the trigger. The second issue is the menu not always
displaying on click, so previous hover logic has been re-used,
which is now possible because the trigger has been refactored
to support menu display on hover.