fix(1-on-1): update tests for displaying videos with toolbar
jitsi-meet will have changes that have the remote videos of
the filmstrip display with the toolbars. The 1-on-1 tests have
to make sure the toolbars are not set to show all the time
and also wait for them to disappear. The test for showing
remote videos while sharing video was removed because sharing
video docks the toolbar, so remote videos always display. That
test has been replaced with a test to ensure remote videos
stay displayed while being hovered over.