ref(WebParticipant): use one way to determine a boolean result
There's another method that just lives next to it, but it return
a Boolean wrapper which can be null and cleaning that up is a different
story(all scripts would have to be checked what they expect - maybe
it's possible to go with primitive who knows). Then that method also
deals with the WebDriver directly which makes it an even longer story.