ref(Participant): mobile does not execute scripts (#143)
* ref(Participant): mobile does not execute scripts

Mobile is not capable of executing any JavaScript code and probably
never will be, so move away any code that depends on that to
the WebParticipant.

* fix(WebParticipant): throw if not JavaScriptExecutor

The methods executing JavaScript should fail instead of silently
returning null.

* fix(WebParticipant.getProtocol): remove unnecessary check

* ref(WebParticipant): use one way to determine a boolean result

There's another method that just lives next to it, but it return
a Boolean wrapper which can be null and cleaning that up is a different
story(all scripts would have to be checked what they expect - maybe
it's possible to go with primitive who knows). Then that method also
deals with the WebDriver directly which makes it an even longer story.