fix(AndroidParticipantOptions): provide default app/package
It turns out that Android driver will not simply accept the package name
as the "app" capability to run the app if it comes pre-installed on
the device (in contrary to what the iOS driver does). Because of that we
need to provide default values for both package and activity
capabilities.

Rename the MobileParticipant's app field to something more meaningful.