fix(MobileParticipant): join not working on uiautomator2
One newer Android phones old automator will not find "conference"
element, after joining the room. When uiautomator2 is used this will
work, but then it inputs the room name another way which does not popup
the keyboard and 'hideKeyboard' will throw an Exception.

This method works in both cases and gets us closer to the iOS scenario.