fix(TestHint): Android is reported either as LINUX or ANDROID
Not sure if it's when running Appium from source or what, but ANDROID is
reported as LINUX and iOS as MAC. It could be also something with the
new appium-java-client, but I really don't want to deal with this right
now. TestHint is used only on mobile anyway, so we can't really mistake
it with a web platform.