fix: Fixes failing lobby test and missing first participant.
Seems like from time to time we quickly hangup one of the participants in testConferenceEndedInLobby test, and it stays in the room with lobby enabled. In that case when participant 1 in next test tries to join fails on isInMuc check as there is Lobby. We close it deciding that chrome crashed and try again. The problem was that our close is wrong and the indexes are messed up and participant2 becomes participant1 and everything starts failing :)
Fixes the close which is used only in that rare case we consider as chrome crash ... we may eventually get rid of that workaround as we no longer see chrome crashes.