fix: Disable chrome debug logging.
Since Chrome 102 we see strange random crashes when running chrome. Reporting to chromium and it seems it is triggered inside `if (VLOG_IS_ON(3))` block. At the moment those logs are being cleared once session is over or chrome crashes, so disabling them for now. Thanks to @JonathanLennox for digging into it.